apps/worker/pyproject.toml
[project]
name = "yng-voice-worker"
version = "0.1.0"
description = "YouNet Voice Studio — inference worker"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "psycopg2-binary>=2.9.10",
    "redis>=5.2.1",
    "httpx>=0.28.0",
    "structlog>=25.1.0",
    "pyloudnorm>=0.1.1",
    "pydub>=0.25.1",
    "cryptography>=45.0.0",
    "soundfile>=0.13.1",
    "librosa>=0.11.0",
    "numpy>=1.26.0",
    "scipy>=1.13.0",
    "PyNaCl>=1.5.0",
    "boto3>=1.35.0",
    "google-generativeai>=0.8.0",
    "openai>=1.59.0",  # elevenlabs client uses openai-compatible or direct REST
    "faster-whisper>=1.1.0",
    "pyannote.audio>=3.3.2",
    "silero-vad>=5.1.2",
    "underthesea>=6.8.4",
    "torch>=2.5.0",
    "torchaudio>=2.5.0",
    "TTS>=0.22.0",
    "prometheus-client>=0.21.1",
    "python-multipart>=0.0.20",
    "opentelemetry-api>=1.28.0",
    "opentelemetry-sdk>=1.28.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.28.0",
]

[project.optional-dependencies]
vieneu = [
    "vieneu>=1.0.0",
]
voxcpm = [
    "voxcpm>=0.2.0",
]
dev = [
    "pytest>=8.3.0",
    "pytest-cov>=6.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
    "respx>=0.21.0",
    "mypy>=1.13.0",
    "ruff>=0.8.0",
]
# Lightweight: only what's needed to run unit tests without heavy ML deps.
# Use with: pip install -e ".[test-only]"
test-only = [
    "pytest>=8.3.0",
    "pytest-cov>=6.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "redis>=5.2.1",
    "structlog>=25.1.0",
    "soundfile>=0.13.1",
    "numpy>=1.26.0",
    "scipy>=1.13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "C4", "SIM", "TRY", "RUF"]
ignore = ["TRY003", "TRY400"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = ["integration: marks integration tests"]

[tool.coverage.run]
source = ["src/worker"]
omit = [
    "src/worker/main.py",
    "src/worker/tracing.py",
    "src/worker/providers/*",
]

[tool.coverage.report]
fail_under = 40
show_missing = true

[tool.hatch.build.targets.wheel]
packages = ["src/worker"]
